+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 Xtreme3D version history
 ========================
+v4.1.0 (by Gecko) - ??/??/2025
+------------------------------
+- INI file functions
+- Audio functions
+- Video functions
+- WindowControl functions
+- `WindowSetBackgroundColor`.
+
 v4.0.0 (by Gecko) - 09/04/2025
 ------------------------------
 - Based on GLScene 2.2 and Delphi 10.4